move a thing
